+++ datalad/support/tests/test_gitrepo.py
@@ -877,69 +877,6 @@ def test_get_added_files_commit_msg():
     eq_(f(["f1", "f2"]), 'Added 2 files\n\nFiles:\nf1\nf2')
 
 
-@with_tempfile(mkdir=True)
-@with_tempfile(mkdir=True)
-def test_git_custom_calls(path, path2):
-    # we need a GitRepo instance
-    repo = GitRepo(path, create=True)
-    with open(op.join(path, "cc_test.dat"), 'w') as f:
-        f.write("test_git_custom_calls")
-
-    out, err = repo._gitpy_custom_call('add', 'cc_test.dat')
-
-    # actually executed:
-    assert_in("cc_test.dat", repo.get_indexed_files())
-    ok_(repo.dirty)
-
-    # call using cmd_options:
-    out, err = repo._gitpy_custom_call('commit',
-                                       cmd_options={'m': 'added file'})
-    ok_clean_git(path, annex=False)
-    # check output:
-    assert_in("1 file changed", out)
-    assert_in("cc_test.dat", out)
-    eq_('', err)
-
-    # impossible 'add' call should raise ...
-    assert_raises(GitCommandError, repo._gitpy_custom_call,
-                  'add', 'not_existing', expect_fail=False)
-    # .. except we expect it to fail:
-    repo._gitpy_custom_call('add', 'not_existing', expect_fail=True)
-
-    # log outputs:
-    with swallow_logs(new_level=logging.DEBUG) as cm:
-        out, err = repo._gitpy_custom_call('status',
-                                           log_stdout=True,
-                                           log_stderr=True)
-
-        assert_in("On branch master", out)
-        assert_in("nothing to commit", out)
-        eq_("", err)
-        for line in out.splitlines():
-            assert_in("stdout| " + line, cm.out)
-
-    # don't log outputs:
-    with swallow_logs(new_level=logging.DEBUG) as cm:
-        out, err = repo._gitpy_custom_call('status',
-                                           log_stdout=False,
-                                           log_stderr=False)
-
-        assert_in("On branch master", out)
-        assert_in("nothing to commit", out)
-        eq_("", err)
-        eq_("", cm.out)
-
-    # use git_options:
-    # Note: 'path2' doesn't contain a git repository
-    with assert_raises(GitCommandError) as cm:
-        repo._gitpy_custom_call('status', git_options={'C': path2})
-    assert_in("-C %s status" % path2, str(cm.exception))
-    assert_re_in("fatal: [Nn]ot a git repository",
-                 str(cm.exception), match=False)
-
-    # TODO: How to test 'env'?
-
-
 @with_testrepos(flavors=['local'])
 @with_tempfile(mkdir=True)
 def test_get_tracking_branch(o_path, c_path):
@@ -1129,7 +1066,8 @@ def test_optimized_cloning(path):
         return dict(
             [(os.path.join(*o.split(os.sep)[-2:]),
               os.stat(o).st_ino)
-             for o in glob(os.path.join(repo.repo.git_dir,
+             for o in glob(os.path.join(repo.path,
+                                        repo.get_git_dir(repo),
                                         'objects', '*', '*'))])
 
     origin_inodes = _get_inodes(repo)
@@ -1208,6 +1146,10 @@ def test_GitRepo_gitignore(path):
 
     gr = GitRepo(path, create=True)
     sub = GitRepo(op.join(path, 'ignore-sub.me'))
+    # we need to commit something, otherwise add_submodule
+    # will already refuse the submodule for having no commit
+    sub.add('a_file.txt')
+    sub.commit()
 
     from ..exceptions import GitIgnoreError
 
@@ -1301,6 +1243,15 @@ def test_gitattributes(path):
             'tag': False,
             'sec.key': 'val',
         })
+
+
+@with_tempfile(mkdir=True)
+def test_get_hexsha_tag(path):
+    gr = GitRepo(path, create=True)
+    gr.commit(msg="msg", options=["--allow-empty"])
+    gr.tag("atag", message="atag msg")
+    # get_hexsha() dereferences a tag to a commit.
+    eq_(gr.get_hexsha("atag"), gr.get_hexsha())
 
 
 @with_tempfile(mkdir=True)